gc + unicodedata + _struct: extract into module/{gc,unicodedata,_struct}/
importing.rs shrinks by 489 lines.
