+++ pyre/pyre-interpreter/src/module/gc/mod.rs
@@ -1,0 +1,4 @@
+//! gc module — PyPy: pypy/module/gc/
+
+pub mod interp_gc;
+pub mod moduledef;
