+++ pyre/pyre-interpreter/src/module/unicodedata/mod.rs
@@ -1,0 +1,7 @@
+//! unicodedata module — PyPy: pypy/module/unicodedata/
+//!
+//! Stub providing normalize() and category() — enough to let
+//! `import unicodedata` succeed.
+
+pub mod interp_unicodedata;
+pub mod moduledef;
