+++ pyre/pyre-interpreter/src/module/unicodedata/moduledef.rs
@@ -1,0 +1,7 @@
+//! unicodedata module definition — PyPy: pypy/module/unicodedata/moduledef.py
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_unicodedata::register_module(ns);
+}
